Fix MANPATH configuration in .profile
Problem: The MANPATH variable must ends with ':' if we want the system
man pages to be availble to the user. The current .profile explicitly
ensure this doesn't happen if there is at least one user man page
directory.

Solution: Correctly update the MANPATH so that if the MANPATH is empty
before adding a directory, the MANPATH ends with ':'.