[autoconf] Fix config.site to take C(XX)FLAGS into account
Problem: when using different C(XX)FLAGS the config.site directed to the
same (unique) cache file. This caused lots of build error because
autotools check the config cache was indeed generated using the same
C(XX)FLAGS.

Solution: take into account the value of C(XX)FLAGS when computing the
name of the cache file. We use a suffix computed using the md5 sum of
the C(XX)FLAGS to differentiate the cache files.