[wip] Start working on a tool to help format code
Problem: When editing code in vim (or emacs) one often wants to format
array data so that it looks good on screen. Unfortunately formating
tabular data is a tedious process.

Solution: Automate with a perl script. One just filter the visual
selection through the perl script and obtain a good looking table.