[codefmt] Add cli options and fix last line formatting
Problem: We need command line options. The last line should not print
the end-of-line marker and we should not output columns separators if
there is no data in the column on the last line of output.

Solution: Add options processing with Getopt::Long Fix the formatting of
the last line using a special case handling.