[setup] Follow PREFIX environment variable
Problem: When preparing a release we want to test the release candidate
without overwriting our home directory files.

Solution: Instead of using '~' and '$HOME' as prefix for all install
paths, use '$PREFIX' and let the user configure this environment
variable before calling the setup script. In case this variable is not
set, use the value of '$HOME'.