[ruby] Add a default .gemrc
Problem: When using ruby, gem installs packages in the global directory
by default. We should instead default to install in a user directory.
Also when installing scripts, gem default to install in ~/bin. We should
instead install in ~/.local/bin as we do for everything.

Solution: create a ~/.gemrc, which is a yaml file, and setup additional
options for the 'gem' and 'install' targets. These additional options
will trigger a user-install and choose ~/.local/bin as script install
directory.