Embed a simple uu-decoder
Problem: Some files embedded in the setup script are uu-encoded. The
script tries to use uudecode to decode them, but on some systems
uudecode is not installed.

Solution: Embed a perl script that can uu-decode its standard input. In
the setup script, check if uudecode is available. It it is not, rename
the perl script to be called uudecode and update the PATH so that it is
available to the setup script.