Update nerd-fonts handling
Problem: The nerd-font installed was not the latest and wasn't a
monospace font. Also the font-size selected by .Xresources.user didn't
take into account the resolution of the display.

Solution: Fix the url of the downloaded font. Add C preprocessor macros
to select the font face size according to the screen Y resolution.