Add a wrapper for cygwin's git to please VS Code
Problem: Visual Studio Code in windows expect file paths returned by
'git' to be in windows format. When using cygwin's git, VS-Code choke
on the returned paths and doesn't display anything (not even an error
message). Also, VS-Code insists that the 'git' command is a real
executable (it doesn't know how to launch shell-scripts).

Solution: Wrap cygwin's git with an executable that will filter the
output and transform paths from cygwin (posix) to windows. The wrapper
knows when to do these conversion when it sees 'rev-parse' as a first
argument.