Found a flaw in .Xresources file configuration
Problem: The .Xresources file configures the XTerm*fontFaceName
resource. This resource specifies the name of the true-type font that
xterm should use. Unfortunately, true-type fonts are rendered on the
client, using fonts installed on the client. This resource should then
be configured on the client, not through the .Xresource which is server
side configuration.

Solution: Add a TODO item to fix this problem later. A possible
implementation is to use a .Xdefaults file.