Removed use of ^? for erase instead of ^H
Problem: ^H is an emacs shortcut and it is also the standard erase
backspace sequence for xterm. In order to solve this issue we configured
xterm to use ^? instead. But this conflicts with vim which uses terminfo
to know what is the erase sequence and thus expect ^H.

Solution: Remove the override in Xresources