Clean scripts to run bash from the new windows terminal
Problem: We have two scripts w/ the same name and they do different
things and too many things. One is for loading VC++ environment in a
%COMSPEC% shell, the second is for running a 'bash' from the windows
terminal. The second script also tries to load the VC++ environment,
but because it runs a 'bash -l' later, part of the environment is
overwritten and thus the VC++ environment is not fully functional.

Solution: Remove vcshell, which spawns a %COMSPEC% w/ a VC++
environment, we can do that directly from windows terminal by running
vcvarsall.bat. Rename vcshell.bat to become bshell.bat and remove the
loading of VC++ environment. We can load the VC++ environment correctly
from within a bash shell using msvc-shell.